+++ lib/Reconciler.lua
@@ -31,6 +31,7 @@ local Change = require(script.Parent.Change)
 local getDefaultPropertyValue = require(script.Parent.getDefaultPropertyValue)
 local SingleEventManager = require(script.Parent.SingleEventManager)
 local Symbol = require(script.Parent.Symbol)
+local GlobalConfig = require(script.Parent.GlobalConfig)
 
 local isInstanceHandle = Symbol.named("isInstanceHandle")
 
@@ -317,6 +318,19 @@ function Reconciler._reconcileInternal(instanceHandle, newElement)
 	-- If the element changes type, we assume its subtree will be substantially
 	-- different. This lets us skip comparisons of a large swath of nodes.
 	if oldElement.component ~= newElement.component then
+		if GlobalConfig.getValue("warnOnTypeChange") then
+			warn(("A Roact component is changing type from %s to %s during reconciliation!\n"
+				.. "This can cause performance issues; see issue #88 for details."):format(
+				tostring(oldElement.component),
+				tostring(newElement.component)
+			))
+
+			print(("Old element source: %s\nNew element source: %s"):format(
+				oldElement.source or DEFAULT_SOURCE,
+				newElement.source or DEFAULT_SOURCE
+			))
+		end
+
 		local parent = instanceHandle._parent
 		local key = instanceHandle._key